Major navbar rework and added profile page
